pyproject.toml
[project]
name = "lovable-clone"
version = "0.1"
description = "Cantata"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.1.0",
    "langchain-google-genai>=3.2.0",
    "langchain-groq>=1.1.0",
    "langgraph>=1.0.3",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "textual>=6.6.0",
    "ujson>=5.11.0",
]
